+++ test/socketio/pcp_socket.io.test.js
@@ -42,6 +42,15 @@ field2:value2
 
 `;
 
+const pcpMessage4 = `pcp-action:ABC
+pcp-event:notification4
+pcp-body-type:text
+action:MESSAGE
+field1:value1
+field2:value2
+
+Header`;
+
 describe("PCP", () => {
   let socket;
 
@@ -57,6 +66,7 @@ describe("PCP", () => {
     const waitNotification1Promise = waitForEvent(socket, "notification1");
     const waitNotification2Promise = waitForEvent(socket, "notification2");
     const waitNotification3Promise = waitForEvent(socket, "notification3");
+    const waitNotification4Promise = waitForEvent(socket, "notification4");
     const result = await emitEvent(socket, "sendNotification", pcpMessage);
     expect(result).toBe(true);
     const waitResult1 = await waitNotification1Promise;
@@ -65,6 +75,8 @@ describe("PCP", () => {
     expect(waitResult2).toEqual(pcpMessage2);
     const waitResult3 = await waitNotification3Promise;
     expect(waitResult3).toEqual(pcpMessage3);
+    const waitResult4 = await waitNotification4Promise;
+    expect(waitResult4).toEqual(pcpMessage4);
   });
 
   test("PCP format error", async () => {